+++ __tools__/clean-capi/big-bullets.js
@@ -1,0 +1,2 @@
+// @flow
+export default s => s.replace(/•/g, '<span class="bullet">&bull;</span>');
