+++ package.json
@@ -11,6 +11,7 @@
     },
     "homepage": "https://github.com/guardian/guui#readme",
     "dependencies": {
+        "compose-function": "^3.0.3",
         "curlyquotes": "^1.3.2",
         "dompurify": "^1.0.3",
         "emotion": "^9.0.2",
